requirements.txt
pytesseract
Pillow
flask
opencv-python
cv2